web_requirements.txt
flask==2.2.3
flask-cors==3.0.10
waitress==2.1.2
python-dotenv==1.0.0 